netty (internal): Migrate test cases from airframe-finagle
